requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
environ==1.0
idna==3.4
Markdown==3.4.1
numpy==1.24.1
pandas==1.5.2
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
scipy==1.10.0
six==1.16.0
sklearn==0.0.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
xgboost==1.7.3
